+++ CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file. This change log follows the conventions
 of [keepachangelog.com](http://keepachangelog.com/).
 
+## 0.2.0 - 2025-01-26
+
+### Fixed
+
+- Job execution is now correctly being stopped when the scheduler component is stopped.
+
+### Added
+
+- Added the compatibility with Pedestal Interceptor while scheduling jobs.
+
 ## 0.1.0 - 2025-01-14
 
 ### Added